Rename Metadata Nodes
