+++ projects/uk.ac.man.cs.llvm/src/uk/ac/man/cs/llvm/ir/model/metadata/MetadataCompositeType.java
@@ -32,15 +32,19 @@ package uk.ac.man.cs.llvm.ir.model.metadata;
 import uk.ac.man.cs.llvm.ir.model.MetadataBlock;
 import uk.ac.man.cs.llvm.ir.model.MetadataBlock.MetadataReference;
 
-public class LocalVariable implements MetadataBaseNode {
+public class MetadataCompositeType implements MetadataBaseNode {
 
     protected MetadataReference context = MetadataBlock.voidRef;
     protected MetadataReference name = MetadataBlock.voidRef;
     protected MetadataReference file = MetadataBlock.voidRef;
     protected long line;
-    protected long arg;
-    protected MetadataReference type = MetadataBlock.voidRef;
+    protected long size;
+    protected long align;
+    protected long offset;
     protected long flags;
+    protected MetadataReference derivedFrom = MetadataBlock.voidRef;
+    protected MetadataReference memberDescriptors = MetadataBlock.voidRef;
+    protected long runtimeLanguage;
 
     public MetadataReference getContext() {
         return context;
@@ -74,20 +78,28 @@ public class LocalVariable implements MetadataBaseNode {
         this.line = line;
     }
 
-    public long getArg() {
-        return arg;
+    public long getSize() {
+        return size;
     }
 
-    public void setArg(long arg) {
-        this.arg = arg;
+    public void setSize(long size) {
+        this.size = size;
     }
 
-    public MetadataReference getType() {
-        return type;
+    public long getAlign() {
+        return align;
     }
 
-    public void setType(MetadataReference type) {
-        this.type = type;
+    public void setAlign(long align) {
+        this.align = align;
+    }
+
+    public long getOffset() {
+        return offset;
+    }
+
+    public void setOffset(long offset) {
+        this.offset = offset;
     }
 
     public long getFlags() {
@@ -98,8 +110,34 @@ public class LocalVariable implements MetadataBaseNode {
         this.flags = flags;
     }
 
+    public MetadataReference getDerivedFrom() {
+        return derivedFrom;
+    }
+
+    public void setDerivedFrom(MetadataReference derivedFrom) {
+        this.derivedFrom = derivedFrom;
+    }
+
+    public MetadataReference getMemberDescriptors() {
+        return memberDescriptors;
+    }
+
+    public void setMemberDescriptors(MetadataReference memberDescriptors) {
+        this.memberDescriptors = memberDescriptors;
+    }
+
+    public long getRuntimeLanguage() {
+        return runtimeLanguage;
+    }
+
+    public void setRuntimeLanguage(long runtimeLanguage) {
+        this.runtimeLanguage = runtimeLanguage;
+    }
+
     @Override
     public String toString() {
-        return "LocalVariable [context=" + context + ", name=" + name + ", file=" + file + ", line=" + line + ", arg=" + arg + ", type=" + type + ", flags=" + flags + "]";
+        return "CompositeType [context=" + context + ", name=" + name + ", file=" + file + ", line=" + line + ", size=" + size + ", align=" + align + ", offset=" + offset + ", flags=" + flags +
+                        ", derivedFrom=" + derivedFrom + ", memberDescriptors=" + memberDescriptors + ", runtimeLanguage=" + runtimeLanguage + "]";
     }
+
 }
