+++ projects/uk.ac.man.cs.llvm/src/uk/ac/man/cs/llvm/ir/model/metadata/MetadataLexicalBlockFile.java
@@ -32,11 +32,9 @@ package uk.ac.man.cs.llvm.ir.model.metadata;
 import uk.ac.man.cs.llvm.ir.model.MetadataBlock;
 import uk.ac.man.cs.llvm.ir.model.MetadataBlock.MetadataReference;
 
-public class LexicalBlock implements MetadataBaseNode {
+public class MetadataLexicalBlockFile implements MetadataBaseNode {
 
     protected MetadataReference file = MetadataBlock.voidRef;
-    protected long line;
-    protected long column;
 
     public MetadataReference getFile() {
         return file;
@@ -46,24 +44,9 @@ public class LexicalBlock implements MetadataBaseNode {
         this.file = file;
     }
 
-    public long getLine() {
-        return line;
-    }
-
-    public void setLine(long line) {
-        this.line = line;
-    }
-
-    public long getColumn() {
-        return column;
-    }
-
-    public void setColumn(long column) {
-        this.column = column;
-    }
-
     @Override
     public String toString() {
-        return "LexicalBlock [file=" + file + ", line=" + line + ", column=" + column + "]";
+        return "LexicalBlockFile [file=" + file + "]";
     }
+
 }
